update: error delete and update

+++ src/useCase/user/delete-user-usecase.ts
@@ -4,6 +4,11 @@ export class DeleteUserUseCase {
   constructor(private userRepository: UserRepository) {}
 
   async execute({ id }: Input): Promise<Output> {
+    const userExist = await this.userRepository.findById(id);
+    if (!userExist) {
+      throw new Error("User does not exist");
+    }
+
     const user = await this.userRepository.delete(id);
     return { user };
   }
